Corrects condition to make Series-6 Controllers work again
Adaptec did a refactoring from 52011 to 60001 where they introduced a bug so that the Series-6 Controllers are not working anymore. Instead of checking for Series 6, 7 and 8 they suddenly only are checking for 7 and 8 before a certain function call. This commit restores the original logic - making sure also the Series-6 Controller gets it init. Change is only one line of Code. Other controllers are not affected by this change.

See https://github.com/xcp-ng/xcp/issues/472

Signed-off-by: Christian Großegger <chr.grossegger@gmail.com>